refactor: extract helpers, remove dead code, consolidate AppState
Dead code removal:
- Remove ManniesResponse (unused struct in types.rs)
- Remove _dist_hint parameter from travel_go_sector and its call sites
- Remove state_requests_quit() trivial wrapper (replaced by state.quit directly)

AppState consolidation:
- Move DETACH_MODES constant from main.rs to app.rs
- Move collect_mineable_candidates and collect_asteroid_candidates from free
  functions in main.rs to AppState methods
- Extract probe_current_sector_scan() private helper — identical inline lookup
  was duplicated 5x across collect_salvage_candidates, collect_deploy_candidates,
  collect_detached_containers, and the two new methods

Navigation:
- Add list_nav(code, sel, count) -> Option<usize> — single source of truth for
  Up/Down/j/k list navigation, replacing 12 copy-pasted blocks across handlers
- Fix scanner hint: [↑↓/jk] history  [JK] scroll

HTTP client:
- Extract send_with_body() — post() and patch() were identical aside from the method

UI overlays:
- Add render_pick_list() helper — 9 pick-list overlay variants collapsed into
  one reusable function (-286 lines in cockpit.rs)

Module split:
- Extract src/input.rs — handle_event, all handle_*_event handlers, list_nav,
  neighbors_d1, face_d2 (main.rs: 1503 → 187 lines)
- Extract src/api/tasks.rs — all fetch_* spawner functions
