+++ src/api/mod.rs
@@ -1,2 +1,3 @@
 pub mod client;
+pub mod tasks;
 pub mod types;
